+++ axelor-account/src/main/java/com/axelor/apps/account/service/PartnerAccountService.java
@@ -1,0 +1,17 @@
+package com.axelor.apps.account.service;
+
+import com.axelor.apps.account.db.FiscalPosition;
+import com.axelor.apps.base.db.Partner;
+
+public class PartnerAccountService {
+
+  public String getDefaultSpecificTaxNote(Partner partner) {
+    FiscalPosition fiscalPosition = partner.getFiscalPosition();
+
+    if (fiscalPosition == null || !fiscalPosition.getCustomerSpecificNote()) {
+      return "";
+    }
+
+    return fiscalPosition.getCustomerSpecificNoteText();
+  }
+}